requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
